+++ opentelemetry-semantic-conventions/src/opentelemetry/semconv/_incubating/attributes/message_attributes.py
@@ -15,28 +15,30 @@
 from enum import Enum
 from typing import Final
 
+from deprecated import deprecated
+
 MESSAGE_COMPRESSED_SIZE: Final = "message.compressed_size"
 """
-Compressed size of the message in bytes.
+Deprecated: Replaced by `rpc.message.compressed_size`.
 """
 
 MESSAGE_ID: Final = "message.id"
 """
-MUST be calculated as two different counters starting from `1` one for sent messages and one for received message.
-Note: This way we guarantee that the values will be consistent between different implementations.
+Deprecated: Replaced by `rpc.message.id`.
 """
 
 MESSAGE_TYPE: Final = "message.type"
 """
-Whether this is a received or sent message.
+Deprecated: Replaced by `rpc.message.type`.
 """
 
 MESSAGE_UNCOMPRESSED_SIZE: Final = "message.uncompressed_size"
 """
-Uncompressed size of the message in bytes.
+Deprecated: Replaced by `rpc.message.uncompressed_size`.
 """
 
 
+@deprecated(reason="The attribute message.type is deprecated - Replaced by `rpc.message.type`")  # type: ignore
 class MessageTypeValues(Enum):
     SENT: Final = "SENT"
     """sent."""